+++ src/js_realm.hpp
@@ -123,9 +123,7 @@ class RealmDelegate : public BindingContext {
         }
 
         ObjectType realm_object = create_object<T, RealmClass<T>>(m_context, new SharedRealm(realm));
-        ValueType arguments[2];
-        arguments[0] = realm_object;
-        arguments[1] = Value::from_string(m_context, notification_name);
+        ValueType arguments[] = {realm_object, Value::from_string(m_context, notification_name)};
 
         std::list<Protected<FunctionType>> notifications_copy(m_notifications);
         for (auto &callback : notifications_copy) {
@@ -148,6 +146,7 @@ class RealmClass : public ClassDefinition<T, SharedRealm, ObservableClass<T>> {
     using FunctionType = typename T::Function;
     using ObjectType = typename T::Object;
     using ValueType = typename T::Value;
+    using Arguments = js::Arguments<T>;
     using String = js::String<T>;
     using Object = js::Object<T>;
     using Value = js::Value<T>;
@@ -157,7 +156,7 @@ class RealmClass : public ClassDefinition<T, SharedRealm, ObservableClass<T>> {
 public:
     using ObjectDefaultsMap = typename Schema<T>::ObjectDefaultsMap;
     using ConstructorMap = typename Schema<T>::ConstructorMap;
-    
+
     using WaitHandler = void(std::error_code);
     using ProgressHandler = void(uint64_t transferred_bytes, uint64_t transferrable_bytes);
 
@@ -165,24 +164,31 @@ public:
     static FunctionType create_constructor(ContextType);
 
     // methods
-    static void objects(ContextType, FunctionType, ObjectType, size_t, const ValueType[], ReturnValue &);
-    static void object_for_primary_key(ContextType, FunctionType, ObjectType, size_t, const ValueType[], ReturnValue &);
-    static void create(ContextType, FunctionType, ObjectType, size_t, const ValueType[], ReturnValue &);
-    static void delete_one(ContextType, FunctionType, ObjectType, size_t, const ValueType[], ReturnValue &);
-    static void delete_all(ContextType, FunctionType, ObjectType, size_t, const ValueType[], ReturnValue &);
-    static void write(ContextType, FunctionType, ObjectType, size_t, const ValueType[], ReturnValue &);
-    static void add_listener(ContextType, FunctionType, ObjectType, size_t, const ValueType[], ReturnValue &);
-    static void wait_for_download_completion(ContextType, FunctionType, ObjectType, size_t, const ValueType[], ReturnValue &);
-    static void remove_listener(ContextType, FunctionType, ObjectType, size_t, const ValueType[], ReturnValue &);
-    static void remove_all_listeners(ContextType, FunctionType, ObjectType, size_t, const ValueType[], ReturnValue &);
-    static void close(ContextType, FunctionType, ObjectType, size_t, const ValueType[], ReturnValue &);
+    static void objects(ContextType, FunctionType, ObjectType, Arguments, ReturnValue &);
+    static void object_for_primary_key(ContextType, FunctionType, ObjectType, Arguments, ReturnValue &);
+    static void create(ContextType, FunctionType, ObjectType, Arguments, ReturnValue &);
+    static void delete_one(ContextType, FunctionType, ObjectType, Arguments, ReturnValue &);
+    static void delete_all(ContextType, FunctionType, ObjectType, Arguments, ReturnValue &);
+    static void write(ContextType, FunctionType, ObjectType, Arguments, ReturnValue &);
+    static void begin_transaction(ContextType, FunctionType, ObjectType, Arguments, ReturnValue&);
+    static void commit_transaction(ContextType, FunctionType, ObjectType, Arguments, ReturnValue&);
+    static void cancel_transaction(ContextType, FunctionType, ObjectType, Arguments, ReturnValue&);
+    static void add_listener(ContextType, FunctionType, ObjectType, Arguments, ReturnValue &);
+    static void wait_for_download_completion(ContextType, FunctionType, ObjectType, Arguments, ReturnValue &);
+    static void remove_listener(ContextType, FunctionType, ObjectType, Arguments, ReturnValue &);
+    static void remove_all_listeners(ContextType, FunctionType, ObjectType, Arguments, ReturnValue &);
+    static void close(ContextType, FunctionType, ObjectType, Arguments, ReturnValue &);
+    static void compact(ContextType, FunctionType, ObjectType, Arguments, ReturnValue &);
+    static void delete_model(ContextType, FunctionType, ObjectType, Arguments, ReturnValue &);
 
     // properties
     static void get_empty(ContextType, ObjectType, ReturnValue &);
     static void get_path(ContextType, ObjectType, ReturnValue &);
     static void get_schema_version(ContextType, ObjectType, ReturnValue &);
     static void get_schema(ContextType, ObjectType, ReturnValue &);
+    static void get_in_memory(ContextType, ObjectType, ReturnValue &);
     static void get_read_only(ContextType, ObjectType, ReturnValue &);
+    static void get_is_in_transaction(ContextType, ObjectType, ReturnValue &);
 #if REALM_ENABLE_SYNC
     static void get_sync_session(ContextType, ObjectType, ReturnValue &);
 #endif
@@ -191,9 +197,10 @@ public:
     static void constructor(ContextType, ObjectType, size_t, const ValueType[]);
     static SharedRealm create_shared_realm(ContextType, realm::Realm::Config, bool, ObjectDefaultsMap &&, ConstructorMap &&);
 
-    static void schema_version(ContextType, FunctionType, ObjectType, size_t, const ValueType[], ReturnValue &);
-    static void clear_test_state(ContextType, FunctionType, ObjectType, size_t, const ValueType[], ReturnValue &);
-    static void copy_bundled_realm_files(ContextType, FunctionType, ObjectType, size_t, const ValueType[], ReturnValue &);
+    static void schema_version(ContextType, FunctionType, ObjectType, Arguments, ReturnValue &);
+    static void clear_test_state(ContextType, FunctionType, ObjectType, Arguments, ReturnValue &);
+    static void copy_bundled_realm_files(ContextType, FunctionType, ObjectType, Arguments, ReturnValue &);
+    static void delete_file(ContextType, FunctionType, ObjectType, Arguments, ReturnValue &);
 
     // static properties
     static void get_default_path(ContextType, ObjectType, ReturnValue &);
@@ -205,6 +212,7 @@ public:
         {"schemaVersion", wrap<schema_version>},
         {"clearTestState", wrap<clear_test_state>},
         {"copyBundledRealmFiles", wrap<copy_bundled_realm_files>},
+        {"deleteFile", wrap<delete_file>},
         {"_waitForDownload", wrap<wait_for_download_completion>},
     };
 
@@ -219,10 +227,15 @@ public:
         {"delete", wrap<delete_one>},
         {"deleteAll", wrap<delete_all>},
         {"write", wrap<write>},
+        {"beginTransaction", wrap<begin_transaction>},
+        {"commitTransaction", wrap<commit_transaction>},
+        {"cancelTransaction", wrap<cancel_transaction>},
         {"addListener", wrap<add_listener>},
         {"removeListener", wrap<remove_listener>},
         {"removeAllListeners", wrap<remove_all_listeners>},
         {"close", wrap<close>},
+        {"compact", wrap<compact>},
+        {"deleteModel", wrap<delete_model>},
     };
 
     PropertyMap<T> const properties = {
@@ -230,7 +243,9 @@ public:
         {"path", {wrap<get_path>, nullptr}},
         {"schemaVersion", {wrap<get_schema_version>, nullptr}},
         {"schema", {wrap<get_schema>, nullptr}},
+        {"inMemory", {wrap<get_in_memory>, nullptr}},
         {"readOnly", {wrap<get_read_only>, nullptr}},
+        {"isInTransaction", {wrap<get_is_in_transaction>, nullptr}},
 #if REALM_ENABLE_SYNC
         {"syncSession", {wrap<get_sync_session>, nullptr}},
 #endif
@@ -367,6 +382,12 @@ void RealmClass<T>::constructor(ContextType ctx, ObjectType this_object, size_t 
                 config.path = js::default_path();
             }
 
+            static const String in_memory_string = "inMemory";
+            ValueType in_memory_value = Object::get_property(ctx, object, in_memory_string);
+            if (!Value::is_undefined(ctx, in_memory_value) && Value::validated_to_boolean(ctx, in_memory_value, "inMemory")) {
+                config.in_memory = true;
+            }
+
             static const String read_only_string = "readOnly";
             ValueType read_only_value = Object::get_property(ctx, object, read_only_string);
             if (!Value::is_undefined(ctx, read_only_value) && Value::validated_to_boolean(ctx, read_only_value, "readOnly")) {
@@ -376,7 +397,7 @@ void RealmClass<T>::constructor(ContextType ctx, ObjectType this_object, size_t 
             static const String schema_string = "schema";
             ValueType schema_value = Object::get_property(ctx, object, schema_string);
             if (!Value::is_undefined(ctx, schema_value)) {
-                ObjectType schema_object = Value::validated_to_object(ctx, schema_value, "schema");
+                ObjectType schema_object = Value::validated_to_array(ctx, schema_value, "schema");
                 config.schema.emplace(Schema<T>::parse_schema(ctx, schema_object, defaults, constructors));
                 schema_updated = true;
             }
@@ -388,6 +409,28 @@ void RealmClass<T>::constructor(ContextType ctx, ObjectType this_object, size_t 
             }
             else if (schema_updated) {
                 config.schema_version = 0;
+            }
+
+            static const String compact_on_launch_string = "shouldCompactOnLaunch";
+            ValueType compact_value = Object::get_property(ctx, object, compact_on_launch_string);
+            if (!Value::is_undefined(ctx, compact_value)) {
+                if (config.schema_mode == SchemaMode::ReadOnly) {
+                    throw std::invalid_argument("Cannot set 'shouldCompactOnLaunch' when 'readOnly' is set.");
+                }
+                if (config.sync_config) {
+                    throw std::invalid_argument("Cannot set 'shouldCompactOnLaunch' when 'sync' is set.");
+                }
+
+                FunctionType should_compact_on_launch_function = Value::validated_to_function(ctx, compact_value, "shouldCompactOnLaunch");
+                config.should_compact_on_launch_function = [=](uint64_t total_bytes, uint64_t unused_bytes) {
+                    ValueType arguments[2] = {
+                        Value::from_number(ctx, total_bytes),
+                        Value::from_number(ctx, unused_bytes)
+                    };
+
+                    ValueType should_compact = Function<T>::callback(ctx, should_compact_on_launch_function, this_object, 2, arguments);
+                    return Value::to_boolean(ctx, should_compact);
+                };
             }
 
             static const String migration_string = "migration";
@@ -460,13 +503,13 @@ SharedRealm RealmClass<T>::create_shared_realm(ContextType ctx, realm::Realm::Co
 }
 
 template<typename T>
-void RealmClass<T>::schema_version(ContextType ctx, FunctionType, ObjectType this_object, size_t argc, const ValueType arguments[], ReturnValue &return_value) {
-    validate_argument_count(argc, 1, 2);
+void RealmClass<T>::schema_version(ContextType ctx, FunctionType, ObjectType this_object, Arguments args, ReturnValue &return_value) {
+    args.validate_maximum(2);
 
     realm::Realm::Config config;
-    config.path = normalize_realm_path(Value::validated_to_string(ctx, arguments[0]));
-    if (argc == 2) {
-        auto encryption_key = Value::validated_to_binary(ctx, arguments[1], "encryptionKey");
+    config.path = normalize_realm_path(Value::validated_to_string(ctx, args[0]));
+    if (args.count == 2) {
+        auto encryption_key = Value::validated_to_binary(ctx, args[1], "encryptionKey");
         config.encryption_key.assign(encryption_key.data(), encryption_key.data() + encryption_key.size());
     }
 
@@ -481,16 +524,57 @@ void RealmClass<T>::schema_version(ContextType ctx, FunctionType, ObjectType thi
 
 
 template<typename T>
-void RealmClass<T>::clear_test_state(ContextType ctx, FunctionType, ObjectType this_object, size_t argc, const ValueType arguments[], ReturnValue &return_value) {
-    validate_argument_count(argc, 0);
-
+void RealmClass<T>::clear_test_state(ContextType ctx, FunctionType, ObjectType this_object, Arguments args, ReturnValue &return_value) {
+    args.validate_maximum(0);
     js::clear_test_state();
 }
 
 template<typename T>
-void RealmClass<T>::copy_bundled_realm_files(ContextType ctx, FunctionType, ObjectType this_object, size_t argc, const ValueType arguments[], ReturnValue &return_value) {
-    validate_argument_count(argc, 0);
+void RealmClass<T>::copy_bundled_realm_files(ContextType ctx, FunctionType, ObjectType this_object, Arguments args, ReturnValue &return_value) {
+    args.validate_maximum(0);
     realm::copy_bundled_realm_files();
+}
+
+template<typename T>
+void RealmClass<T>::delete_file(ContextType ctx, FunctionType, ObjectType this_object, Arguments args, ReturnValue &return_value) {
+    args.validate_maximum(1);
+
+    ValueType value = args[0];
+    if (!Value::is_object(ctx, value)) {
+        throw std::runtime_error("Invalid argument, expected a Realm configuration object");
+    }
+
+    ObjectType object = Value::validated_to_object(ctx, value);
+    realm::Realm::Config config;
+
+    static const String path_string = "path";
+    ValueType path_value = Object::get_property(ctx, object, path_string);
+    if (!Value::is_undefined(ctx, path_value)) {
+        config.path = Value::validated_to_string(ctx, path_value, "path");
+    }
+    else if (config.path.empty()) {
+        config.path = js::default_path();
+    }
+
+    config.path = normalize_realm_path(config.path);
+
+    std::string realm_file_path = config.path;
+    realm::remove_file(realm_file_path);
+    realm::remove_file(realm_file_path + ".lock");
+    realm::remove_file(realm_file_path + ".note");
+    realm::remove_directory(realm_file_path + ".management");
+
+}
+
+template<typename T>
+void RealmClass<T>::delete_model(ContextType ctx, FunctionType, ObjectType this_object, Arguments args, ReturnValue &return_value) {
+    args.validate_maximum(1);
+    ValueType value = args[0];
+
+    SharedRealm& realm = *get_internal<T, RealmClass<T>>(this_object);
+
+    std::string model_name = Value::validated_to_string(ctx, value, "deleteModel");
+    ObjectStore::delete_data_for_object(realm->read_group(), model_name);
 }
 
 template<typename T>
@@ -529,8 +613,18 @@ void RealmClass<T>::get_schema(ContextType ctx, ObjectType object, ReturnValue &
 }
 
 template<typename T>
+void RealmClass<T>::get_in_memory(ContextType ctx, ObjectType object, ReturnValue &return_value) {
+    return_value.set(get_internal<T, RealmClass<T>>(object)->get()->config().in_memory);
+}
+
+template<typename T>
 void RealmClass<T>::get_read_only(ContextType ctx, ObjectType object, ReturnValue &return_value) {
     return_value.set(get_internal<T, RealmClass<T>>(object)->get()->config().read_only());
+}
+
+template<typename T>
+void RealmClass<T>::get_is_in_transaction(ContextType ctx, ObjectType object, ReturnValue &return_value) {
+    return_value.set(get_internal<T, RealmClass<T>>(object)->get()->is_in_transaction());
 }
 
 #if REALM_ENABLE_SYNC
@@ -547,12 +641,17 @@ void RealmClass<T>::get_sync_session(ContextType ctx, ObjectType object, ReturnV
 #endif
 
 template<typename T>
-void RealmClass<T>::wait_for_download_completion(ContextType ctx, FunctionType, ObjectType this_object, size_t argc, const ValueType arguments[], ReturnValue &return_value) {
-    validate_argument_count(argc, 2);
-    auto callback_function = Value::validated_to_function(ctx, arguments[1]);
+void RealmClass<T>::wait_for_download_completion(ContextType ctx, FunctionType, ObjectType this_object, Arguments args, ReturnValue &return_value) {
+    args.validate_maximum(3);
+    auto callback_function = Value::validated_to_function(ctx, args[1 + (args.count == 3)]);
+
+    ValueType session_callback = Value::from_null(ctx);
+    if (args.count == 3) {
+        session_callback = Value::validated_to_function(ctx, args[1]);
+    }
 
 #if REALM_ENABLE_SYNC
-    auto config_object = Value::validated_to_object(ctx, arguments[0]);
+    auto config_object = Value::validated_to_object(ctx, args[0]);
     ValueType sync_config_value = Object::get_property(ctx, config_object, "sync");
     if (!Value::is_undefined(ctx, sync_config_value)) {
         realm::Realm::Config config;
@@ -563,14 +662,14 @@ void RealmClass<T>::wait_for_download_completion(ContextType ctx, FunctionType, 
             auto encryption_key = Value::validated_to_binary(ctx, encryption_key_value, "encryptionKey");
             config.encryption_key.assign(encryption_key.data(), encryption_key.data() + encryption_key.size());
         }
-        
+
         Protected<ObjectType> thiz(ctx, this_object);
         SyncClass<T>::populate_sync_config(ctx, thiz, config_object, config);
 
         Protected<FunctionType> protected_callback(ctx, callback_function);
         Protected<ObjectType> protected_this(ctx, this_object);
         Protected<typename T::GlobalContext> protected_ctx(Context<T>::get_global_context(ctx));
-        
+
         EventLoopDispatcher<WaitHandler> wait_handler([=](std::error_code error_code) {
             HANDLESCOPE
             if (!error_code) {
@@ -590,7 +689,7 @@ void RealmClass<T>::wait_for_download_completion(ContextType ctx, FunctionType, 
         });
         std::function<WaitHandler> waitFunc = std::move(wait_handler);
 
-        std::function<ProgressHandler> progressFunc; 
+        std::function<ProgressHandler> progressFunc;
 
         auto realm = realm::Realm::get_shared_realm(config);
         if (auto sync_config = config.sync_config)
@@ -623,10 +722,18 @@ void RealmClass<T>::wait_for_download_completion(ContextType ctx, FunctionType, 
             std::shared_ptr<SyncUser> user = sync_config->user;
             if (user && user->state() != SyncUser::State::Error) {
                 if (auto session = user->session_for_on_disk_path(config.path)) {
+                    if (!Value::is_null(ctx, session_callback)) {
+                        FunctionType session_callback_func = Value::to_function(ctx, session_callback);
+                        auto syncSession = create_object<T, SessionClass<T>>(ctx, new WeakSession(session));
+                        ValueType callback_arguments[1];
+                        callback_arguments[0] = syncSession;
+                        Function<T>::callback(protected_ctx, session_callback_func, protected_this, 1, callback_arguments);
+                    }
+
                     if (progressFuncDefined) {
                         session->register_progress_notifier(std::move(progressFunc), SyncSession::NotifierType::download, false);
-                    } 
-                    
+                    }
+
                     session->wait_for_download_completion([=](std::error_code error_code) {
                         realm->close(); //capture and keep realm instance for until here
                         waitFunc(error_code);
@@ -651,25 +758,25 @@ void RealmClass<T>::wait_for_download_completion(ContextType ctx, FunctionType, 
 }
 
 template<typename T>
-void RealmClass<T>::objects(ContextType ctx, FunctionType, ObjectType this_object, size_t argc, const ValueType arguments[], ReturnValue &return_value) {
-    validate_argument_count(argc, 1);
+void RealmClass<T>::objects(ContextType ctx, FunctionType, ObjectType this_object, Arguments args, ReturnValue &return_value) {
+    args.validate_maximum(1);
 
     SharedRealm realm = *get_internal<T, RealmClass<T>>(this_object);
     std::string object_type;
-    validated_object_schema_for_value(ctx, realm, arguments[0], object_type);
+    validated_object_schema_for_value(ctx, realm, args[0], object_type);
 
     return_value.set(ResultsClass<T>::create_instance(ctx, realm, object_type));
 }
 
 template<typename T>
-void RealmClass<T>::object_for_primary_key(ContextType ctx, FunctionType, ObjectType this_object, size_t argc, const ValueType arguments[], ReturnValue &return_value) {
-    validate_argument_count(argc, 2);
+void RealmClass<T>::object_for_primary_key(ContextType ctx, FunctionType, ObjectType this_object, Arguments args, ReturnValue &return_value) {
+    args.validate_maximum(2);
 
     SharedRealm realm = *get_internal<T, RealmClass<T>>(this_object);
     std::string object_type;
-    auto &object_schema = validated_object_schema_for_value(ctx, realm, arguments[0], object_type);
+    auto &object_schema = validated_object_schema_for_value(ctx, realm, args[0], object_type);
     NativeAccessor accessor(ctx, realm, object_schema);
-    auto realm_object = realm::Object::get_for_primary_key(accessor, realm, object_schema, arguments[1]);
+    auto realm_object = realm::Object::get_for_primary_key(accessor, realm, object_schema, args[1]);
 
     if (realm_object.is_valid()) {
         return_value.set(RealmObjectClass<T>::create_instance(ctx, std::move(realm_object)));
@@ -680,21 +787,22 @@ void RealmClass<T>::object_for_primary_key(ContextType ctx, FunctionType, Object
 }
 
 template<typename T>
-void RealmClass<T>::create(ContextType ctx, FunctionType, ObjectType this_object, size_t argc, const ValueType arguments[], ReturnValue &return_value) {
-    validate_argument_count(argc, 2, 3);
+void RealmClass<T>::create(ContextType ctx, FunctionType, ObjectType this_object, Arguments args, ReturnValue &return_value) {
+    args.validate_maximum(3);
 
     SharedRealm realm = *get_internal<T, RealmClass<T>>(this_object);
+    realm->verify_open();
     std::string object_type;
-    auto &object_schema = validated_object_schema_for_value(ctx, realm, arguments[0], object_type);
+    auto &object_schema = validated_object_schema_for_value(ctx, realm, args[0], object_type);
 
-    ObjectType object = Value::validated_to_object(ctx, arguments[1], "properties");
-    if (Value::is_array(ctx, arguments[1])) {
+    ObjectType object = Value::validated_to_object(ctx, args[1], "properties");
+    if (Value::is_array(ctx, args[1])) {
         object = Schema<T>::dict_for_property_array(ctx, object_schema, object);
     }
 
     bool update = false;
-    if (argc == 3) {
-        update = Value::validated_to_boolean(ctx, arguments[2], "update");
+    if (args.count == 3) {
+        update = Value::validated_to_boolean(ctx, args[2], "update");
     }
 
     NativeAccessor accessor(ctx, realm, object_schema);
@@ -703,15 +811,16 @@ void RealmClass<T>::create(ContextType ctx, FunctionType, ObjectType this_object
 }
 
 template<typename T>
-void RealmClass<T>::delete_one(ContextType ctx, FunctionType, ObjectType this_object, size_t argc, const ValueType arguments[], ReturnValue &return_value) {
-    validate_argument_count(argc, 1);
+void RealmClass<T>::delete_one(ContextType ctx, FunctionType, ObjectType this_object, Arguments args, ReturnValue &return_value) {
+    args.validate_maximum(1);
 
     SharedRealm realm = *get_internal<T, RealmClass<T>>(this_object);
+    realm->verify_open();
     if (!realm->is_in_transaction()) {
         throw std::runtime_error("Can only delete objects within a transaction.");
     }
 
-    ObjectType arg = Value::validated_to_object(ctx, arguments[0]);
+    ObjectType arg = Value::validated_to_object(ctx, args[0], "object");
 
     if (Object::template is_instance<RealmObjectClass<T>>(ctx, arg)) {
         auto object = get_internal<T, RealmObjectClass<T>>(arg);
@@ -750,10 +859,11 @@ void RealmClass<T>::delete_one(ContextType ctx, FunctionType, ObjectType this_ob
 }
 
 template<typename T>
-void RealmClass<T>::delete_all(ContextType ctx, FunctionType, ObjectType this_object, size_t argc, const ValueType arguments[], ReturnValue &return_value) {
-    validate_argument_count(argc, 0);
+void RealmClass<T>::delete_all(ContextType ctx, FunctionType, ObjectType this_object, Arguments args, ReturnValue &return_value) {
+    args.validate_maximum(0);
 
     SharedRealm realm = *get_internal<T, RealmClass<T>>(this_object);
+    realm->verify_open();
 
     if (!realm->is_in_transaction()) {
         throw std::runtime_error("Can only delete objects within a transaction.");
@@ -765,18 +875,18 @@ void RealmClass<T>::delete_all(ContextType ctx, FunctionType, ObjectType this_ob
 }
 
 template<typename T>
-void RealmClass<T>::write(ContextType ctx, FunctionType, ObjectType this_object, size_t argc, const ValueType arguments[], ReturnValue &return_value) {
-    validate_argument_count(argc, 1);
+void RealmClass<T>::write(ContextType ctx, FunctionType, ObjectType this_object, Arguments args, ReturnValue &return_value) {
+    args.validate_maximum(1);
 
     SharedRealm realm = *get_internal<T, RealmClass<T>>(this_object);
-    FunctionType callback = Value::validated_to_function(ctx, arguments[0]);
+    FunctionType callback = Value::validated_to_function(ctx, args[0]);
 
     realm->begin_transaction();
 
     try {
         Function<T>::call(ctx, callback, this_object, 0, nullptr);
     }
-    catch (std::exception &e) {
+    catch (...) {
         realm->cancel_transaction();
         throw;
     }
@@ -785,53 +895,83 @@ void RealmClass<T>::write(ContextType ctx, FunctionType, ObjectType this_object,
 }
 
 template<typename T>
-void RealmClass<T>::add_listener(ContextType ctx, FunctionType, ObjectType this_object, size_t argc, const ValueType arguments[], ReturnValue &return_value) {
-    validate_argument_count(argc, 2);
-
-    validated_notification_name(ctx, arguments[0]);
-    auto callback = Value::validated_to_function(ctx, arguments[1]);
+void RealmClass<T>::begin_transaction(ContextType ctx, FunctionType, ObjectType this_object, Arguments args, ReturnValue &return_value) {
+    args.validate_maximum(0);
 
     SharedRealm realm = *get_internal<T, RealmClass<T>>(this_object);
-    if (realm->is_closed()) {
-        throw ClosedRealmException();
-    }
+    realm->begin_transaction();
+}
+
+template<typename T>
+void RealmClass<T>::commit_transaction(ContextType ctx, FunctionType, ObjectType this_object, Arguments args, ReturnValue &return_value) {
+    args.validate_maximum(0);
+
+    SharedRealm realm = *get_internal<T, RealmClass<T>>(this_object);
+    realm->commit_transaction();
+}
+
+template<typename T>
+void RealmClass<T>::cancel_transaction(ContextType ctx, FunctionType, ObjectType this_object, Arguments args, ReturnValue &return_value) {
+    args.validate_maximum(0);
+
+    SharedRealm realm = *get_internal<T, RealmClass<T>>(this_object);
+    realm->cancel_transaction();
+}
+
+template<typename T>
+void RealmClass<T>::add_listener(ContextType ctx, FunctionType, ObjectType this_object, Arguments args, ReturnValue &return_value) {
+    args.validate_maximum(2);
+
+    validated_notification_name(ctx, args[0]);
+    auto callback = Value::validated_to_function(ctx, args[1]);
+
+    SharedRealm realm = *get_internal<T, RealmClass<T>>(this_object);
+    realm->verify_open();
     get_delegate<T>(realm.get())->add_notification(callback);
 }
 
 template<typename T>
-void RealmClass<T>::remove_listener(ContextType ctx, FunctionType, ObjectType this_object, size_t argc, const ValueType arguments[], ReturnValue &return_value) {
-    validate_argument_count(argc, 2);
+void RealmClass<T>::remove_listener(ContextType ctx, FunctionType, ObjectType this_object, Arguments args, ReturnValue &return_value) {
+    args.validate_maximum(2);
 
-    validated_notification_name(ctx, arguments[0]);
-    auto callback = Value::validated_to_function(ctx, arguments[1]);
+    validated_notification_name(ctx, args[0]);
+    auto callback = Value::validated_to_function(ctx, args[1]);
 
     SharedRealm realm = *get_internal<T, RealmClass<T>>(this_object);
-    if (realm->is_closed()) {
-        throw ClosedRealmException();
-    }
+    realm->verify_open();
     get_delegate<T>(realm.get())->remove_notification(callback);
 }
 
 template<typename T>
-void RealmClass<T>::remove_all_listeners(ContextType ctx, FunctionType, ObjectType this_object, size_t argc, const ValueType arguments[], ReturnValue &return_value) {
-    validate_argument_count(argc, 0, 1);
-    if (argc) {
-        validated_notification_name(ctx, arguments[0]);
+void RealmClass<T>::remove_all_listeners(ContextType ctx, FunctionType, ObjectType this_object, Arguments args, ReturnValue &return_value) {
+    args.validate_maximum(1);
+    if (args.count) {
+        validated_notification_name(ctx, args[0]);
     }
 
     SharedRealm realm = *get_internal<T, RealmClass<T>>(this_object);
-    if (realm->is_closed()) {
-        throw ClosedRealmException();
-    }
+    realm->verify_open();
     get_delegate<T>(realm.get())->remove_all_notifications();
 }
 
 template<typename T>
-void RealmClass<T>::close(ContextType ctx, FunctionType, ObjectType this_object, size_t argc, const ValueType arguments[], ReturnValue &return_value) {
-    validate_argument_count(argc, 0);
+void RealmClass<T>::close(ContextType ctx, FunctionType, ObjectType this_object, Arguments args, ReturnValue &return_value) {
+    args.validate_maximum(0);
 
     SharedRealm realm = *get_internal<T, RealmClass<T>>(this_object);
     realm->close();
+}
+
+template<typename T>
+void RealmClass<T>::compact(ContextType ctx, FunctionType, ObjectType this_object, Arguments args, ReturnValue &return_value) {
+    args.validate_maximum(0);
+
+    SharedRealm realm = *get_internal<T, RealmClass<T>>(this_object);
+    if (realm->is_in_transaction()) {
+        throw std::runtime_error("Cannot compact a Realm within a transaction.");
+    }
+
+    return_value.set(realm->compact());
 }
 
 } // js